Clean-up dp

+++ include/imm/seq_table.h
@@ -2,6 +2,7 @@
 #define IMM_SEQ_TABLE_H
 
 #include "imm/export.h"
+#include <inttypes.h>
 
 struct imm_abc;
 struct imm_seq;
@@ -13,10 +14,9 @@ IMM_EXPORT void                  imm_seq_table_destroy(struct imm_seq_table cons
 IMM_EXPORT int    imm_seq_table_add(struct imm_seq_table* table, struct imm_seq const* seq,
                                     double lprob);
 IMM_EXPORT int    imm_seq_table_normalize(struct imm_seq_table* table);
-IMM_EXPORT double imm_seq_table_lprob(struct imm_seq_table const* table,
-                                      struct imm_seq const*       seq);
+IMM_EXPORT double imm_seq_table_lprob(struct imm_seq_table const* table, struct imm_seq const* seq);
 IMM_EXPORT struct imm_abc const* imm_seq_table_get_abc(struct imm_seq_table const* table);
-IMM_EXPORT unsigned              imm_seq_table_min_seq(struct imm_seq_table const* table);
-IMM_EXPORT unsigned              imm_seq_table_max_seq(struct imm_seq_table const* table);
+IMM_EXPORT uint8_t               imm_seq_table_min_seq(struct imm_seq_table const* table);
+IMM_EXPORT uint8_t               imm_seq_table_max_seq(struct imm_seq_table const* table);
 
 #endif
